Refactor Kins component into Showcase; update imports and styles
