+++ rokawoo-react-website/src/App.tsx
@@ -5,7 +5,7 @@ import styles from "./App.module.css"
 import { Navbar } from "./components/Navbar/Navbar";
 import { Hero } from './components/Hero/Hero';
 import { About } from './components/About/About';
-import { Kins } from './components/Kins/Kins';
+import { Kins } from './components/Showcase/Showcase';
 import { Projects } from './components/Projects/Projects';
 import { Contact} from './components/Contact/Contact';
