+++ rokawoo-react-website/src/components/Showcase/scripts/spin-animation.ts
